fix: FAQ 문구 수정 / 폰트 main -> Main으로 변경
