+++ src/components/attendance/AttendanceButton.tsx
@@ -11,7 +11,7 @@ const AttendanceButton = () => {
   return (
     <>
       <div
-        className='px-4 py-3 cursor-pointer text-center text-[16px] font-bold font-main bg-primary300 text-white rounded-xl'
+        className='px-4 py-3 cursor-pointer text-center text-[16px] font-bold font-Main bg-primary300 text-white rounded-xl'
         onClick={openAttendanceModal}
       >
         참여의사 전달하기
